Convert to build as a module using hatchling
Converts the project to be structured so that it can be built into
wheels using `python3 -m build`. The build system chosen is hatchling,
because it is very simple to use.

Note that now instead of using ./main.py to invoke the script, the
module installs a command called `md-spec-parser` that is used instead

Also, the name of the project is changed to "md-spec-parser" since
"spec-parser" is already taken in pip