rdf: Use @vocab for global NamedIndividuals
Changes the JSON Schema so that properties that reference classes use
@vocab instead of @id for their type. The difference is that @vocab will
treat the value as a term and expand it using the current context (in
this case, just the global context). The practical side effect of this
is that @vocab will correctly examine if the value is a global named
individual, whereas @id will not.

As an example:
```json
{
  "@context": {
    "class-prop": {
      "@id": "http://example.com/class-prop",
      "@type": "@vocab"
    },
    "named-individual": "http://example.com/named-individual"
  },
  "@id": "_:blank",
  "class-prop": "named-individual"
}
```
will expand the class-prop value to
"http://example.com/named-individual" where as:

```json
{
  "@context": {
    "class-prop": {
      "@id": "http://example.com/class-prop",
      "@type": "@id"
    },
    "named-individual": "http://example.com/named-individual"
  },
  "@id": "_:blank",
  "class-prop": "named-individual"
}
```

will not.