Fix named individuals in JSON-LD context
The "global" named individuals (e.g. NoneElement) were not added to the
JSON-LD context, likely because the heuristic used to identify them was
broken. Fix this, but use a different mechanism that finds all vocab
values values, and anything that is not a vocab is considered to be a
"global" named individual.