Split ontology and shapes graph
Having a unified graph presents a problem for SHACL validation because
instances of objects (e.g. Named Individuals) in the shapes graph are
validated against the shapes. This doesn't work because the Named
Individuals are not fully defined objects, just references.

The solution to this is to split the shapes graph and the ontology graph
into separate files. SHACL validation tools allow a separate ontology
file to be specified which is not subject to the shape graph validation.